Health bar adjustments
Moved the health bar widget set up to be in the player BP instead of the enemy BP so we can keep things relating to the player and things relating to the enemy separate.

Also made some visual adjustments to the health bar and divided by max health so the current health value properly displays on the health bar.